requirements.txt
asgiref==3.5.2
Django==4.0.6
django-environ==0.9.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
marshmallow==3.17.0
packaging==21.3
psycopg2==2.9.3
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
